+++ ComAxpShop/app/Http/Controllers/orderApiController.php
@@ -15,71 +15,85 @@ use App\Models\Product;
 class orderApiController extends Controller
 {
     public function getAllOrder(){
-        $orders = DB::table('orders')
-                        ->join('orderdetails', 'orders.orderNumber', '=', 'orderdetails.orderNumber')
-                        ->orderBy('orders.orderNumber', 'asc')
-                        ->get();
-        return $orders;
+        $user = auth()->user();
+        
+        if($user->tokenCan('Employee')){  
+            $orders = DB::table('orders')
+                            ->join('orderdetails', 'orders.orderNumber', '=', 'orderdetails.orderNumber')
+                            ->orderBy('orders.orderNumber', 'asc')
+                            ->get();
+            return $orders;
+        }
+        else{
+            return response()->json(['errors' => 'you have no permission to access this page'], 401);
+        }
     }
 
     public function addOrder(Request $request){
-        $products = $request->only(['productCode', 'quantityOrdered', 'priceEach', 'orderLineNumber']);
-        $input = $request->except(['productCode', 'quantityOrdered', 'priceEach', 'orderLineNumber']);
-
-        $validStatus = ['canceled', 'disputed', 'in process', 'on hold', 'resloved','shipped'];
-
-        // validate
+        $user = auth()->user();
         
-        $productValidator = $this->productValidator($products);
+        if($user->tokenCan('Salesman')){  
+            $products = $request->only(['productCode', 'quantityOrdered', 'priceEach', 'orderLineNumber']);
+            $input = $request->except(['productCode', 'quantityOrdered', 'priceEach', 'orderLineNumber']);
 
-        $generalValidator = Validator::make($input, [
-            'requiredDate' => 'required|date_format:Y-m-d|after_or_equal:today',
-            'shippedDate' => 'date_format:Y-m-d|after_or_equal:today|before_or_equal:requiredDate',
-            'status' => 'in:'.implode(',', $validStatus),
-            'comments' => 'nullable',
-            'customerNumber' => 'required|exists:customers,customerNumber',
-            'discountCode' => 'nullable|exists:discountcodes,discountCode',
-        ]);
+            $validStatus = ['canceled', 'disputed', 'in process', 'on hold', 'resloved','shipped'];
 
-        if($generalValidator->fails()){
-            return response()->json(['errors' => $generalValidator->errors()], 400);
+            // validate
+            
+            $productValidator = $this->productValidator($products);
+
+            $generalValidator = Validator::make($input, [
+                'requiredDate' => 'required|date_format:Y-m-d|after_or_equal:today',
+                'shippedDate' => 'date_format:Y-m-d|after_or_equal:today|before_or_equal:requiredDate',
+                'status' => 'in:'.implode(',', $validStatus),
+                'comments' => 'nullable',
+                'customerNumber' => 'required|exists:customers,customerNumber',
+                'discountCode' => 'nullable|exists:discountcodes,discountCode',
+            ]);
+
+            if($generalValidator->fails()){
+                return response()->json(['errors' => $generalValidator->errors()], 400);
+            }
+
+            if($productValidator->fails){
+                return response()->json($productValidator->errors, 400);
+            }
+
+            // order
+
+            $dateToday = Carbon\Carbon::now()->setTimezone('Asia/Phnom_Penh')->format('Y-m-d');
+
+            $order = $request->only(['requiredDate', 'shippedDate', 'status', 
+                                    'comments', 'customerNumber', 'discountCode']);
+            $order['orderDate'] = $dateToday;
+            $fetch = $this->addOrderToDB($order);
+
+            // // order detail and calculate gained point
+
+            $productArr = explode(',', $products['productCode']);
+            $quantityArr = explode(',', $products['quantityOrdered']);
+            $priceArr = explode(',', $products['priceEach']);
+            $orderLineArr = explode(',', $products['orderLineNumber']);
+            $productCount = count($productArr);
+
+            $orderNumber = $fetch['orderNumber'];
+
+            for($i = 0; $i < $productCount; $i++){
+                $orderDetail = [
+                    'orderNumber' => $orderNumber,
+                    'productCode' => $productArr[$i],
+                    'quantityOrdered' => $quantityArr[$i],
+                    'priceEach' => $priceArr[$i],
+                    'orderLineNumber' => $orderLineArr[$i],
+                ];
+                $this->addOrderDetailToDB($orderDetail);
+            }
+
+            return response()->json(['message' => 'add order success']);
         }
-
-        if($productValidator->fails){
-            return response()->json($productValidator->errors, 400);
+        else{
+            return response()->json(['errors' => 'you have no permission to access this page'], 401);
         }
-
-        // order
-
-        $dateToday = Carbon\Carbon::now()->setTimezone('Asia/Phnom_Penh')->format('Y-m-d');
-
-        $order = $request->only(['requiredDate', 'shippedDate', 'status', 
-                                'comments', 'customerNumber', 'discountCode']);
-        $order['orderDate'] = $dateToday;
-        $fetch = $this->addOrderToDB($order);
-
-        // // order detail and calculate gained point
-
-        $productArr = explode(',', $products['productCode']);
-        $quantityArr = explode(',', $products['quantityOrdered']);
-        $priceArr = explode(',', $products['priceEach']);
-        $orderLineArr = explode(',', $products['orderLineNumber']);
-        $productCount = count($productArr);
-
-        $orderNumber = $fetch['orderNumber'];
-
-        for($i = 0; $i < $productCount; $i++){
-            $orderDetail = [
-                'orderNumber' => $orderNumber,
-                'productCode' => $productArr[$i],
-                'quantityOrdered' => $quantityArr[$i],
-                'priceEach' => $priceArr[$i],
-                'orderLineNumber' => $orderLineArr[$i],
-            ];
-            $this->addOrderDetailToDB($orderDetail);
-        }
-
-        return response()->json(['message' => 'add order success']);
     }
 
     public function addOrderToDB($orderData){
@@ -104,40 +118,56 @@ class orderApiController extends Controller
         ]);
     }
 
-    public function deleteOrder(Request $request){
-        $validator = Validator::make(request()->all(), [
-            'orderNumber' => 'required|exists:orders,orderNumber',
-        ]);
+    public function deleteOrder($orderNumber){
+        $user = auth()->user();
+        
+        if($user->tokenCan('Salesman')){  
+            $inputArr = ['orderNumber' => $orderNumber];
 
-        if($validator->fails()){
-            return response()->json(['errors' => $validator->errors()], 401);
+            $validator = Validator::make($inputArr, [
+                'orderNumber' => 'required|exists:orders,orderNumber',
+            ]);
+
+            if($validator->fails()){
+                return response()->json(['errors' => $validator->errors()], 401);
+            }
+
+            $targetOrder = Order::where('orderNumber', $orderNumber)->first();
+            $targetOrder->delete();
+
+            return response()->json(['message' => 'delete order successfully']);
         }
-
-        $input = $request->only('orderNumber');
-        $targetOrder = Order::where('orderNumber', $input)->first();
-        $targetOrder->delete();
-
-        return response('delete completed');
+        else{
+            return response()->json(['errors' => 'you have no permission to access this page'], 401);
+        }
     }
 
-    public function updateOrder(Request $request){
-        $validStatus = ['canceled', 'disputed', 'in process', 'on hold', 'resloved', 'shipped'];
+    public function updateOrder($orderNumber, Request $request){
+        $user = auth()->user();
+        
+        if($user->tokenCan('Salesman')){  
+            $validStatus = ['canceled', 'disputed', 'in process', 'on hold', 'resloved', 'shipped'];
 
-        $validator = Validator::make(request()->all(), [
-            'orderNumber' => 'required|exists:orders,orderNumber',
-            'status' => 'in:'.implode(',', $validStatus),
-        ]);
+            $input = $request->all();
+            $input['orderNumber'] = $orderNumber;
 
-        if($validator->fails()){
-            return response()->json(['errors' => $validator->errors()], 401);
+            $validator = Validator::make($input, [
+                'orderNumber' => 'required|exists:orders,orderNumber',
+                'status' => 'in:'.implode(',', $validStatus),
+            ]);
+
+            if($validator->fails()){
+                return response()->json(['errors' => $validator->errors()], 401);
+            }
+
+            $targerOrder = Order::find($orderNumber);
+            $targerOrder->fill($input)->update();
+
+            return response()->json(['message' => 'update order successfully']);
         }
-
-        $input = $request->all();
-
-        $targerOrder = Order::find($input['orderNumber']);
-        $targerOrder->fill($input)->update();
-
-        return response('Data updated');
+        else{
+            return response()->json(['errors' => 'you have no permission to access this page'], 401);
+        }
     }
 
     public function productValidator($products){
@@ -173,7 +203,7 @@ class orderApiController extends Controller
                 return $returnedObject;
             }
 
-            $isAvailable = $this->isAvailable($product);
+            $isAvailable = $this->isProductAvailable($product);
             if(!$isAvailable){
                 $returnedObject->fails = true;
                 $returnedObject->errors = ['quantityOrdered' => "Product amount is not enough for purchase"];
@@ -183,7 +213,7 @@ class orderApiController extends Controller
         return $returnedObject;
     }
 
-    public function isAvailable($product){
+    public function isProductAvailable($product){
         $targetProduct = Product::find($product['productCode']);
 
         if($targetProduct->quantityInStock - $product['quantityOrdered'] < 0){
