+++ ComAxpShop/app/Http/Controllers/productApiController.php
@@ -10,13 +10,13 @@ use Validator;
 
 class productApiController extends Controller
 {
-    public function getProduct(Request $request){
+    public function getCatalog(Request $request){
         $type = $request['type'];
 
         switch($type){
             case 0: // show all PRODUCT
                 $products = DB::table('products')
-                            ->select('productName', 'buyPrice')
+                            ->select('productName', 'buyPrice', 'quantityInStock')
                             ->orderBy('productName', 'asc')
                             ->get();
                 return $products;
@@ -24,7 +24,7 @@ class productApiController extends Controller
             case 1: // show by PRODUCT VENDOR
                 $productVendor = $request['productVendor'];
                 $products = DB::table('products')
-                            ->select('productName', 'buyPrice')
+                            ->select('productName', 'buyPrice', 'quantityInStock')
                             ->where('productVendor', $productVendor)
                             ->orderBy('productVendor', 'asc')
                             ->get();
@@ -33,7 +33,7 @@ class productApiController extends Controller
             case 2: // show by PRODUCT SCALE
                 $productScale = $request['productScale'];
                 $products = DB::table('products')
-                            ->select('productName', 'buyPrice')
+                            ->select('productName', 'buyPrice', 'quantityInStock')
                             ->where('productScale', $productScale)
                             ->orderBy('productVendor', 'asc')
                             ->get();
@@ -42,7 +42,7 @@ class productApiController extends Controller
             case 3: // show by PRODUCT LINE
                 $productLine = $request['productLine'];
                 $products = DB::table('products')
-                            ->select('productName', 'buyPrice')
+                            ->select('productName', 'buyPrice', 'quantityInStock')
                             ->where('productLine', $productLine)
                             ->orderBy('productLine', 'asc')
                             ->get();
@@ -50,27 +50,50 @@ class productApiController extends Controller
         }
     }
 
-    public function addProduct(Request $request){
-        $validator = Validator::make(request()->all(), [
-            'productCode' => 'required|min:8|unique:products',
-            'productName' => 'required',
-            'productLine' => 'required|exists:productlines,productLine',
-            'productScale' => 'required' ,
-            'productVendor' => 'required',
-            'productDescription' => 'required',
-            'quantityInStock' => 'required',
-            'buyPrice' => 'required',
-            'MSRP' => 'required'
-        ]);
+    public function getProduct(){
 
-        if($validator->fails()){
-            return response()->json(['errors' => $validator->errors()], 400);
+        $user = auth()->user();
+
+        if($user->tokenCan('Employee')){
+            $products = DB::table('products')
+                        ->orderBy('productCode', 'asc')
+                        ->get();
+
+            return $products;
         }
-        $product = $request->all();
+        else{
+            return response()->json(['errors' => 'you have no permission to access this page'], 401);
+        }
+    }
 
-        $fetch = $this->addToDB($product);
-    
-        return response()->json(['message' => 'add product success']);
+    public function addProduct(Request $request){
+        $user = auth()->user();
+        
+        if($user->tokenCan('Salesman')){
+            $validator = Validator::make(request()->all(), [
+                'productCode' => 'required|min:8|unique:products',
+                'productName' => 'required',
+                'productLine' => 'required|exists:productlines,productLine',
+                'productScale' => 'required' ,
+                'productVendor' => 'required',
+                'productDescription' => 'required',
+                'quantityInStock' => 'required',
+                'buyPrice' => 'required',
+                'MSRP' => 'required'
+            ]);
+
+            if($validator->fails()){
+                return response()->json(['errors' => $validator->errors()], 400);
+            }
+            $product = $request->all();
+
+            $fetch = $this->addToDB($product);
+        
+            return response()->json(['message' => 'add product success']);
+        }
+        else{
+            return response()->json(['errors' => 'you have no permission to access this page'], 401);
+        }
     }
 
     public function addToDB($productData){
@@ -87,43 +110,61 @@ class productApiController extends Controller
           ]);
     }
 
-    public function deleteProduct(Request $request){
-        $validator = Validator::make(request()->all(), [
-            'productCode' => 'required|exists:products,productCode',
-        ]);
+    public function deleteProduct($productCode){
+        $user = auth()->user();
+        if($user->tokenCan('Salesman')){
 
-        if($validator->fails()){
-            return response()->json(['errors' => $validator->errors()], 400);
+            $input = ['productCode' => $productCode];
+
+            $validator = Validator::make($input, [
+                'productCode' => 'required|exists:products,productCode',
+            ]);
+
+            if($validator->fails()){
+                return response()->json(['errors' => $validator->errors()], 400);
+            }
+
+            $targetProduct = Product::where('productCode', $productCode)->first();
+            $targetProduct->delete();
+
+            return response()->json(['message' => 'delete product success']);
         }
-
-        $input = $request->only('productCode');
-        $targetProduct = Product::where('productCode', $input)->first();
-        $targetProduct->delete();
-
-        return response()->json(['message' => 'delete product success']);
+        else{
+            return response()->json(['errors' => 'you have no permission to access this page'], 401);
+        }
     }
 
-    public function updateProduct(Request $request){ // product code is not allowed to change
+    public function updateProduct($productCode, Request $request){ // product code is not allowed to change
+        $user = auth()->user();
+        if($user->tokenCan('Salesman')){
+            $input = $request->all();
+            $input['productCode'] = $productCode;
 
-        $validator = Validator::make(request()->all(), [
-            'productCode' => 'required|exists:products,productCode',
-            'productLine' => 'exists:productlines,productLine',
-        ]);
+            $validator = Validator::make($input, [
+                'productCode' => 'required|exists:products,productCode',
+                'productLine' => 'exists:productlines,productLine',
+            ]);
 
-        if($validator->fails()){
-            return response()->json(['errors' => $validator->errors()], 400);
+            if($validator->fails()){
+                return response()->json(['errors' => $validator->errors()], 400);
+            }
+
+            $targetProduct = Product::find($productCode);
+            $targetProduct->fill($input)->update();
+
+            return response()->json(['message' => 'update product success']);
         }
-
-        $input = $request->all();
-
-        $targetProduct = Product::find($input['productCode']);
-        $targetProduct->fill($input)->update();
-
-        return response()->json(['message' => 'update product success']);
+        else{
+            return response()->json(['errors' => 'you have no permission to access this page'], 401);
+        }
     }
 
     static function productSold($product){
-        
+        $targetProduct = Product::find($product['productCode']);
+        $targetProduct['quantityInStock'] -= $product['quantityOrdered'];
+        $targetProduct->update();
+
+        return;
     }
 }
